+++ src/debugServer.ts
@@ -1,4 +1,4 @@
-import { DebugSession, Event, TerminatedEvent } from 'vscode-debugadapter';
+import { DebugSession, Event, TerminatedEvent, Thread, ThreadEvent, StoppedEvent, ContinuedEvent } from 'vscode-debugadapter';
 import { DebugProtocol } from 'vscode-debugprotocol';
 
 /**
@@ -12,6 +12,8 @@ export interface LaunchRequestArguments extends DebugProtocol.LaunchRequestArgum
     /** Run in terminal instead of output pane. */
     interactiveTerminal: boolean;
 }
+
+const THREAD_ID = 0;
 
 export class Ev3devBrowserDebugSession extends DebugSession {
     protected initializeRequest(response: DebugProtocol.InitializeResponse,
@@ -29,6 +31,10 @@ export class Ev3devBrowserDebugSession extends DebugSession {
 
     protected customRequest(command: string, response: DebugProtocol.Response, args: any): void {
         switch (command) {
+            case 'ev3devBrowser.debugger.thread':
+                this.sendEvent(new ThreadEvent(args, THREAD_ID));
+                this.sendResponse(response);
+                break;
             case 'ev3devBrowser.debugger.terminate':
                 this.sendEvent(new TerminatedEvent());
                 this.sendResponse(response);
@@ -39,6 +45,20 @@ export class Ev3devBrowserDebugSession extends DebugSession {
     protected disconnectRequest(response: DebugProtocol.DisconnectResponse,
         args: DebugProtocol.DisconnectArguments): void {
         this.sendEvent(new Event('ev3devBrowser.debugger.stop', args));
+        this.sendResponse(response);
+    }
+
+    protected threadsRequest(response: DebugProtocol.ThreadsResponse): void {
+        response.body = {
+            threads: [
+                new Thread(THREAD_ID, 'thread')
+            ]
+        };
+        this.sendResponse(response);
+    }
+
+    protected pauseRequest(response: DebugProtocol.PauseResponse, args: DebugProtocol.PauseArguments): void {
+        this.sendEvent(new Event('ev3devBrowser.debugger.interrupt', args));
         this.sendResponse(response);
     }
 }